+++ example/lib/main.dart
@@ -1,14 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_colorpicker/flutter_colorpicker.dart';
+import 'package:flutter_colorpicker/material_picker.dart';
 
 main() => runApp(
       MaterialApp(
-        home: Scaffold(
-          appBar: AppBar(
-            title: const Text('Flutter Color Picker Example'),
-          ),
-          body: MyApp(),
-        ),
+        title: 'Flutter Example',
+        theme: ThemeData(primaryColor: Colors.blue),
+        home: MyApp(),
       ),
     );
 
@@ -18,67 +16,92 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  Color pickerColor = Color(0xff443a49);
   Color currentColor = Color(0xff443a49);
 
   ValueChanged<Color> onColorChanged;
 
-  changeColor(Color color) {
-    setState(() {
-      pickerColor = color;
-    });
-  }
+  changeSimpleColor(Color color) => setState(() => currentColor = color);
+  changeMaterialColor(Color color) => setState(() {
+        currentColor = color;
+        Navigator.of(context).pop();
+      });
 
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: RaisedButton(
-        elevation: 3.0,
-        onPressed: () {
-          pickerColor = currentColor;
-          showDialog(
-            context: context,
-            builder: (BuildContext context) {
-              return AlertDialog(
-                titlePadding: EdgeInsets.all(0.0),
-                contentPadding: EdgeInsets.all(0.0),
-                title: Container(
-                  color: Theme.of(context).primaryColor,
-                  child: Center(
-                    child: const Text(
-                      'Pick a color!',
-                      style: TextStyle(
-                        fontSize: 20.0,
-                        color: const Color(0xffffffff),
-                      ),
-                    ),
-                  ),
-                  padding: EdgeInsets.all(20.0),
-                ),
-                content: SingleChildScrollView(
-                  child: ColorPicker(
-                    pickerColor: pickerColor,
-                    onColorChanged: changeColor,
-                    colorPickerWidth: 1000.0,
-                    pickerAreaHeightPercent: 0.7,
-                  ),
-                ),
-                actions: <Widget>[
-                  FlatButton(
-                    child: Text('Got it'),
-                    onPressed: () {
-                      setState(() => currentColor = pickerColor);
-                      Navigator.of(context).pop();
+    return DefaultTabController(
+      length: 2,
+      child: Scaffold(
+        appBar: AppBar(
+          title: Text('Flutter Color Picker Example'),
+          bottom: TabBar(
+            tabs: <Widget>[
+              Tab(text: 'Simple picker'),
+              Tab(text: 'Material picker'),
+            ],
+          ),
+        ),
+        body: TabBarView(
+          children: <Widget>[
+            Center(
+              child: RaisedButton(
+                elevation: 3.0,
+                onPressed: () {
+                  showDialog(
+                    context: context,
+                    builder: (BuildContext context) {
+                      return AlertDialog(
+                        titlePadding: EdgeInsets.all(0.0),
+                        contentPadding: EdgeInsets.all(0.0),
+                        content: SingleChildScrollView(
+                          child: ColorPicker(
+                            pickerColor: currentColor,
+                            onColorChanged: changeSimpleColor,
+                            colorPickerWidth: 1000.0,
+                            pickerAreaHeightPercent: 0.7,
+                            enableAlpha: true,
+                          ),
+                        ),
+                      );
                     },
-                  ),
-                ],
-              );
-            },
-          );
-        },
-        child: Text('Change me'),
-        color: currentColor,
-        textColor: const Color(0xffffffff),
+                  );
+                },
+                child: Text('Change me'),
+                color: currentColor,
+                textColor: useWhiteForeground(currentColor)
+                    ? const Color(0xffffffff)
+                    : const Color(0xff000000),
+              ),
+            ),
+            Center(
+              child: RaisedButton(
+                elevation: 3.0,
+                onPressed: () {
+                  showDialog(
+                    context: context,
+                    builder: (BuildContext context) {
+                      return AlertDialog(
+                        titlePadding: EdgeInsets.all(0.0),
+                        contentPadding: EdgeInsets.all(0.0),
+                        content: SingleChildScrollView(
+                          child: MaterialPicker(
+                            pickerColor: currentColor,
+                            onColorChanged: changeMaterialColor,
+                            enableLabel: true,
+                          ),
+                        ),
+                      );
+                    },
+                  );
+                },
+                child: Text('Change me'),
+                color: currentColor,
+                textColor: useWhiteForeground(currentColor)
+                    ? const Color(0xffffffff)
+                    : const Color(0xff000000),
+              ),
+            ),
+          ],
+        ),
       ),
     );
   }